refactor: 패키지 구조 변경
- enum 클래스를 domain으로 옮김
