+++ src/main/java/oncall/domain/Week.java
@@ -1,4 +1,4 @@
-package oncall.constant;
+package oncall.domain;
 
 public enum Week {
